requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.0
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
pyzbar==0.1.9
opencv-python-headless==4.8.0.74
